requirements.txt
cryptography>=44.0.0
javaproperties>=0.8.0
pydantic>=2.10.0
pyyaml>=6.0.0